Refactor managers
