+++ Assets/Scripts/Managers/BoardManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e0b29f809ca8d9f42a7bf0c1b54044f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
