+++ Assets/Scripts/Managers/PieceManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5d208fe82adfa5d4ba894a8841741152
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
